fold into version_resolver
